Struct pcsc::Card

pub struct Card { /* private fields */ }

A connection to a smart card.

This structure wraps SCARDHANDLE.

Implementations

impl Card

pub fn transaction(&mut self) -> Result<Transaction<'_>, Error>

Start a new exclusive transaction with the card.

Operations on the card for the duration of the transaction can only be performed through the returned Transaction.

This function wraps SCardBeginTransaction (pcsclite, MSDN).

pub fn transaction2(&mut self) -> Result<Transaction<'_>, (&mut Self, Error)>

Start a new exclusive transaction with the card.

Operations on the card for the duration of the transaction can only be performed through the returned Transaction.

This function is like Card::transaction, but also returns the reference to self on error. When starting a transaction, you might want to deal with transient errors, like Error::ResetCard, by reconnecting to the card, and retrying the transaction. When this functionality is wrapped, this doesn’t work, because mutable references can’t be reborrowed (at least in current Rust). This function returns the reference, which allows this construct.

This function wraps SCardBeginTransaction (pcsclite, MSDN).

pub fn reconnect( &mut self, share_mode: ShareMode, preferred_protocols: Protocols, initialization: Disposition ) -> Result<(), Error>

Reconnect to the card.

This function wraps SCardReconnect (pcsclite, MSDN).

pub fn disconnect(self, disposition: Disposition) -> Result<(), (Card, Error)>

Disconnect from the card.

In case of error, ownership of the card is returned to the caller.

This function wraps SCardDisconnect (pcsclite, MSDN).

Note

Card implements Drop which automatically disconnects the card using Disposition::ResetCard; you only need to call this function if you want to handle errors or use a different disposition method.

pub fn status(&self) -> Result<(Status, Protocol), Error>

👎Deprecated since 2.3.0: Use status2() or status2_owned() instead.

Get current info on the card.

This function wraps SCardStatus (pcsclite, MSDN).

Deprecated

The reader names and ATR return values are missing.

When there is no active protocol (as when connecting to the reader directly), this function panics.

Use status2() or status2_owned() instead.

pub fn status2<'names_buf, 'atr_buf>( &self, names_buffer: &'names_buf mut [u8], atr_buffer: &'atr_buf mut [u8] ) -> Result<CardStatus<'names_buf, 'atr_buf>, Error>

Get current info on the card.

names_buffer is a buffer that should be large enough to hold all of the reader names.

atr_buffer is a buffer that should be large enough to hold the ATR. The recommended size is MAX_ATR_SIZE, which should be always sufficent.

The function status2_len can be used to find the exact required lengths.

If the buffers are not large enough to hold all of the names or the ATR, Error::InsufficientBuffer is returned.

This function wraps SCardStatus (pcsclite, MSDN).

pub fn status2_len(&self) -> Result<(usize, usize), Error>

Get the needed length of the names buffer (first result) and ATR buffer (second result) to be passed to status2.

This function wraps SCardStatus (pcsclite, MSDN).

pub fn status2_owned(&self) -> Result<CardStatusOwned, Error>

Get current info on the card, allocating buffers of the required size.

This function wraps SCardStatus (pcsclite, MSDN). It is an owned version of status2, calling status2_len to determine the required buffer sizes.

pub fn get_attribute<'buf>( &self, attribute: Attribute, buffer: &'buf mut [u8] ) -> Result<&'buf [u8], Error>

Get an attribute of the card or card reader.

buffer is a buffer that should be large enough for the attribute data. The function get_attribute_len can be used to find the exact required length.

Returns a slice into buffer containing the attribute data.

If the buffer is not large enough, Error::InsufficientBuffer is returned.

This function wraps SCardGetAttrib (pcsclite, MSDN).

pub fn get_attribute_len(&self, attribute: Attribute) -> Result<usize, Error>

Get the needed length of a buffer to be passed to get_attribute.

This function wraps SCardGetAttrib (pcsclite, MSDN).

pub fn get_attribute_owned( &self, attribute: Attribute ) -> Result<Vec<u8>, Error>

Get an attribute of the card or card reader, allocating a buffer of the required size.

This function wraps SCardGetAttrib (pcsclite, MSDN). It is an owned version of get_attribute, calling get_attribute_len to determine the required buffer size.

pub fn set_attribute( &self, attribute: Attribute, attribute_data: &[u8] ) -> Result<(), Error>

Set an attribute of the card or card reader.

This function wraps SCardSetAttrib (pcsclite, MSDN).

pub fn transmit<'buf>( &self, send_buffer: &[u8], receive_buffer: &'buf mut [u8] ) -> Result<&'buf [u8], Error>

Transmit an APDU command to the card.

receive_buffer is a buffer that should be large enough to hold the APDU response.

Returns a slice into receive_buffer containing the APDU response.

If receive_buffer is not large enough to hold the APDU response, Error::InsufficientBuffer is returned.

This function wraps SCardTransmit (pcsclite, MSDN).

pub fn transmit2<'buf>( &self, send_buffer: &[u8], receive_buffer: &'buf mut [u8] ) -> Result<&'buf [u8], (Error, usize)>

Transmit an APDU command to the card.

This functions works like transmit but the error type is (Error, usize).

receive_buffer is a buffer that should be large enough to hold the APDU response.

Returns a slice into receive_buffer containing the APDU response.

If receive_buffer is not large enough to hold the APDU response, Error::InsufficientBuffer is returned, and the usize value is set to the required size.

usize value of the error has no meaning for other Error values than Error::InsufficientBuffer.

Note that when Error::InsufficientBuffer is returned, the provided command has been already effectively executed by the card. Do not treat this as a generic way to obtain the size of the response as you may end up issuing commands multiple times which can lead to unexpected results. Normally, most operations on standard card let you know the expected size of the response in advance. Be sure to only use this for commands that may be executed multiple times in a row without changing the state of the card.

This function wraps SCardTransmit (pcsclite, MSDN).

pub fn control<'buf>( &self, control_code: DWORD, send_buffer: &[u8], receive_buffer: &'buf mut [u8] ) -> Result<&'buf [u8], Error>

Sends a command directly to the reader (driver).

control_code is the reader-specific control code. You may need to pass it through the ctl_code() function, according to the driver documentation.

receive_buffer is a buffer that should be large enough to hold the response.

Returns a slice into receive_buffer containing the response.

If receive_buffer is not large enough to hold the response, Error::InsufficientBuffer is returned.

This function wraps SCardControl (pcsclite, MSDN).