Struct pcsc::Transaction

pub struct Transaction<'tx> { /* private fields */ }

An exclusive transaction with a card.

A transaction ensures uninterrupted access to the card for its duration. All other operations performed on the same underlying card (even from other processes) will block until the transaction is finished.

Implementations

impl<'tx> Transaction<'tx>

pub fn end( self, disposition: Disposition ) -> Result<(), (Transaction<'tx>, Error)>

End the transaction.

In case of error, ownership of the transaction is returned to the caller.

This function wraps SCardEndTransaction (pcsclite, MSDN).

Note

Transaction implements Drop which automatically ends the transaction using Disposition::LeaveCard; you only need to call this function if you want to handle errors or use a different disposition method.